Be smart about not redoing work.
Don't transcode or copy files when the dest
already exists and is not older than the source.

Testing this is a bit tricky. I'm not as happy with this as with the
tests for get_transcode_filters, because mocking out getmtime feels
like a kludge. I'd rather pass in values than pass in a function.